+++ src/index.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+import TodoProvider from './context/todoProvider';
 
 ReactDOM.render(
     <React.StrictMode>
-        <App />
+        <TodoProvider>
+            <App />
+        </TodoProvider>
     </React.StrictMode>,
     document.getElementById('root')
 );
